🎨: App 컴포넌트 추가 - ui 렌더링 로직 분리
