+++ src/basic/app.js
@@ -1,0 +1,23 @@
+import Cart from './components/cart/Cart';
+import CartAddButton from './components/cartAddButton/CartAddButton';
+import CartTotal from './components/cartTotal/CartTotal';
+import Container from './components/Container';
+import ContentWrapper from './components/ContentWrapper';
+import Header from './components/Header';
+import ProductSelector from './components/productSelector/ProductSelector';
+import StockStatus from './components/stockStatus/StockStatus';
+
+export default function App() {
+  const contentWrapper = ContentWrapper();
+  contentWrapper.appendChild(Header({ title: '장바구니' }));
+  contentWrapper.appendChild(Cart());
+  contentWrapper.appendChild(CartTotal());
+  contentWrapper.appendChild(ProductSelector());
+  contentWrapper.appendChild(CartAddButton());
+  contentWrapper.appendChild(StockStatus());
+
+  const containerDiv = Container();
+  containerDiv.appendChild(contentWrapper);
+
+  return containerDiv;
+}
