+++ src/basic/main.basic.js
@@ -1,37 +1,15 @@
-import Cart from './components/cart/Cart';
-import CartAddButton from './components/cartAddButton/CartAddButton';
-import CartTotal from './components/cartTotal/CartTotal';
+import App from './app';
 import { renderCalculateCart } from './components/cartTotal/renderCalculateCart';
-import Container from './components/Container';
-import ContentWrapper from './components/ContentWrapper';
-import Header from './components/Header';
 import { updateSelectOptions } from './components/productSelect/updateSelectOptions';
-import ProductSelector from './components/productSelector/ProductSelector';
-import StockStatus from './components/stockStatus/StockStatus';
 import { CONSTANTS } from './constants';
 import { ProductStore } from './store/productStore';
 import { helper } from './utils/helper';
 
 function main() {
-  const contentWrapper = ContentWrapper();
-
-  contentWrapper.appendChild(Header({ title: '장바구니' }));
-  contentWrapper.appendChild(Cart());
-  contentWrapper.appendChild(CartTotal());
-  contentWrapper.appendChild(ProductSelector());
-  contentWrapper.appendChild(CartAddButton());
-  contentWrapper.appendChild(StockStatus());
-
-  const containerDiv = Container();
-  containerDiv.appendChild(contentWrapper);
-
   const root = document.getElementById('app');
-  root.appendChild(containerDiv);
+  root.appendChild(App());
 
-  const productStore = ProductStore.getInstance();
-  const products = productStore.getState().products;
-  const lastSelectedItem = productStore.getState().lastSelectedItem;
-
+  const { products, lastSelectedItem } = ProductStore.getInstance().getState();
   updateSelectOptions(products);
   renderCalculateCart(products);
